+++ modules/auth/controller.js
@@ -18,6 +18,29 @@ class AuthController {
             response.error(req, res, error.message, 500);
         }
     }
+
+    async sendMail(req, res){
+        try{
+            const { email } = req.body;
+
+            await service.sendToken(email);
+            response.success(req, res, 'Mail sent', 200);
+        }catch(error){
+            response.error(req, res, error.message, 500);
+        }
+    }
+
+    async updatePassword(req, res){
+        try{
+            const { token, password } = req.body;
+
+            const auth = await service.changePassword(token, password);
+            response.success(req, res, auth.message, 200);
+        }catch(error) {
+            response.error(req, res, error.message, 500);
+        }
+    }
+
 }
 
 module.exports = AuthController;