+++ modules/auth/router.js
@@ -5,5 +5,6 @@ const Controller = new AuthController();
 const router = express.Router();
 
 router.post('/login', (req, res) => { Controller.loginUser(req, res) });
-
+router.post('/forgot-password', (req, res) => { Controller.sendMail(req, res) });
+router.put('/change-password', (req, res) => { Controller.updatePassword(req, res) });
 module.exports = router;