+++ modules/user/router.js
@@ -4,6 +4,6 @@ const UserController = require('./controller');
 const Controller = new UserController();
 const router = express.Router();
 
-router.post('/', (req, res) => { Controller.registerUer(req, res) });
+router.post('/', (req, res) => { Controller.registerUser(req, res) });
 
 module.exports = router;