+++ routes/router.js
@@ -2,10 +2,12 @@ const express = require('express');
 const userRouter = require('../modules/user/router');
 const profileRotuer = require('../modules/profile/router');
 const authRotuer = require('../modules/auth/router');
+const followerRouter = require('../modules/follower/router');
 const router = express.Router();
 
 router.use('/api/v1/users', userRouter);
 router.use('/api/v1/profile', profileRotuer);
 router.use('/api/v1/auth', authRotuer);
+router.use('/api/v1/follower', followerRouter);
 
 module.exports = router;