+++ changelog.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 Keep me updated 🤦.
 
+## 2022-02-28
+
+### Added
+
+- endpoints for follow and unfollow user
+- flow to send mail for recovery password and flow for update password
+- flow to create user with profile and auth values.
+
+### Removed
+
+- following model and following module
+- migratios for alter table auth
+
 ## 2022-02-26
 
 ### Added